+++ tests/unittest/utils/test_functional.py
@@ -124,7 +124,7 @@ class TestAlivenessChecker:
         d.satellite.id = "SAT"
         d.satellite._is_alive
         assert functional.check_aliveness_checkers(d, log_failure=True) is False
-        d.satellite.log_warning.assert_called_with("failed is_living check")
+        d.satellite.logger.warning.assert_called_with("failed is_living check")
 
 
 @pytest.mark.parametrize("prop_name,expected", [("prop", True), ("not_a_prop", False)])